+++ QUANTAXIS/QAFetch/QAQuery.py
@@ -144,6 +144,9 @@ def QA_fetch_stock_list(collections=DATABASE.stock_list):
     '获取股票列表'
     return [item for item in collections.find()]
 
+def QA_fetch_index_list(collections=DATABASE.index_list):
+    '获取指数列表'
+    return [item for item in collections.find()]
 
 def QA_fetch_stock_terminated(collections=DATABASE.stock_terminated):
     '获取股票基本信息 , 已经退市的股票列表'
@@ -250,7 +253,7 @@ def QA_fetch_index_day(code, start, end, format='numpy', collections=DATABASE.in
         for item in cursor:
 
             __data.append([str(item['code']), float(item['open']), float(item['high']), float(
-                item['low']), float(item['close']), float(item['vol']), item['date']])
+                item['low']), float(item['close']), int(item['up_count']), int(item['down_count']), float(item['vol']), float(item['amount']), item['date']])
 
         # 多种数据格式
         if format in ['n', 'N', 'numpy']:
@@ -259,7 +262,7 @@ def QA_fetch_index_day(code, start, end, format='numpy', collections=DATABASE.in
             __data = __data
         elif format in ['P', 'p', 'pandas', 'pd']:
             __data = DataFrame(
-                __data, columns=['code', 'open', 'high', 'low', 'close', 'volume', 'date'])
+                __data, columns=['code', 'open', 'high', 'low', 'close', 'up_count', 'down_count', 'volume', 'amount', 'date'])
             __data['date'] = pd.to_datetime(__data['date'])
             __data = __data.set_index('date', drop=False)
         else:
@@ -299,10 +302,10 @@ def QA_fetch_index_min(
     for item in cursor:
 
         __data.append([str(item['code']), float(item['open']), float(item['high']), float(
-            item['low']), float(item['close']), float(item['vol']), item['datetime'], item['time_stamp'], item['date']])
+            item['low']), float(item['close']), int(item['up_count']), int(item['down_count']), float(item['vol']), float(item['amount']), item['datetime'], item['time_stamp'], item['date']])
 
     __data = DataFrame(__data, columns=[
-        'code', 'open', 'high', 'low', 'close', 'volume', 'datetime', 'time_stamp', 'date'])
+        'code', 'open', 'high', 'low', 'close', 'up_count', 'down_count', 'volume', 'amount', 'datetime', 'time_stamp', 'date'])
 
     __data['datetime'] = pd.to_datetime(__data['datetime'])
     __data = __data.set_index('datetime', drop=False)
@@ -321,6 +324,10 @@ def QA_fetch_future_day():
 def QA_fetch_future_min():
     raise NotImplementedError
 
+
+def QA_fetch_future_list(collections=DATABASE.future_list):
+    '获取期货列表'
+    return [item for item in collections.find()]
 
 def QA_fetch_future_tick():
     raise NotImplementedError