+++ QUANTAXIS/QAMarket/QADealer.py
@@ -79,7 +79,6 @@ class QA_Dealer():
 
     def __init__(self, *args, **kwargs):
         self.deal_name = ''
-        self.deal_engine = {'0x01': self.backtest_stock_dealer}
         self.session = {}
         self.order = None
         self.market_data = None
@@ -98,14 +97,14 @@ class QA_Dealer():
         self.deal_price = 0
         self.deal_amount = 0
         self.order.tax_coeff = order.tax_coeff
-        if order.market_type == MARKET_TYPE.STOCK_CN:
+        # if order.market_type == MARKET_TYPE.STOCK_CN:
 
-            res = self.backtest_stock_dealer()
-            # print(res)
-            self.deal_message[self.order.order_id] = res
+        res = self.backtest_dealer()
+        # print(res)
+        self.deal_message[self.order.order_id] = res
 
-        elif order.market_type == MARKET_TYPE.FUTURE_CN:
-            return self.backtest_future_dealer()
+        # elif order.market_type == MARKET_TYPE.FUTURE_CN:
+        #     return self.backtest_future_dealer()
 
     @property
     def deal_df(self):
@@ -153,7 +152,7 @@ class QA_Dealer():
 
             self.tax = 0  # 买入不收印花税
 
-    def backtest_stock_dealer(self):
+    def backtest_dealer(self):
         # 新增一个__commission_fee_coeff 手续费系数
         """MARKET ENGINE STOCK
 
@@ -167,7 +166,7 @@ class QA_Dealer():
         """
         try:
             if float(self.market_data.get('open')) == float(self.market_data.get('high')) == float(self.market_data.get('close')) == float(self.market_data.get('low')) and \
-                    self.market_data.get('volume') < 4*self.order.amount:
+                    self.market_data.get('volume',self.market_data.get('position')) < 4*self.order.amount:
                 # 调整 : 分钟线 经常处于一个价位 但不代表不能交易 所以加入量的判断(但是不能影响市场, 所以加上4倍量限制)
 
                 self.status = TRADE_STATUS.PRICE_LIMIT
@@ -179,11 +178,11 @@ class QA_Dealer():
                     float(self.order.price) == float(self.market_data.get('low')) or
                     float(self.order.price) == float(self.market_data.get('high'))):
                 '能成功交易的情况 有滑点调整'
-                if float(self.order.amount) < float(self.market_data.get('volume')) * 100 / 16:
+                if float(self.order.amount) < float(self.market_data.get('volume',self.market_data.get('position'))) * 100 / 16:
                     self.deal_price = self.order.price
                     self.deal_amount = self.order.amount
-                elif float(self.order.amount) >= float(self.market_data.get('volume')) * 100 / 16 and \
-                        float(self.order.amount) < float(self.market_data.get('volume')) * 100 / 8:
+                elif float(self.order.amount) >= float(self.market_data.get('volume',self.market_data.get('position'))) * 100 / 16 and \
+                        float(self.order.amount) < float(self.market_data.get('volume',self.market_data.get('position'))) * 100 / 8:
                     """
                     add some slippers
 
@@ -200,36 +199,31 @@ class QA_Dealer():
 
                 else:
                     self.deal_amount = float(
-                        self.market_data.get('volume')) / 8
+                        self.market_data.get('volume',self.market_data.get('position'))) / 8
                     if int(self.order.towards) > 0:
                         self.deal_price = float(self.market_data.get('high'))
                     else:
                         self.deal_price = float(self.market_data.get('low'))
                 self.status = TRADE_STATUS.SUCCESS
                 # print(self.market_data)
-                self.trade_time = self.market_data.get('datetime', self.market_data.get('date',None))
+                self.trade_time = self.market_data.get(
+                    'datetime', self.market_data.get('date', None))
             else:
+                print('failed to deal this order')
+                print(self.order.price)
+                print(self.market_data)
                 self.status = TRADE_STATUS.FAILED
                 self.deal_price = 0
                 self.deal_amount = 0
 
             self.cal_fee()
-            print(self.callback_message)
+            # print(self.callback_message)
             return self.callback_message
 
         except Exception as e:
             QA_util_log_info('MARKET ENGINE ERROR: {}'.format(e))
             self.status = TRADE_STATUS.NO_MARKET_DATA
             return self.callback_message
-
-    def backtest_future_dealer(self):
-        raise NotImplementedError
-
-
-class Stock_Dealer(QA_Dealer):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
 
 if __name__ == '__main__':
     pass